feat: segment the schemas for ease of access
